Fixed that odd 0 word showing up, turns out you need to write at least 4 DAC words to waveform fifo

+++ c_source/cmd_handler.c
@@ -259,7 +259,7 @@ void cmd_handle_command()
 				else if(bytes_to_receive == 8)
 				{
 					//Reassemble the 64-bit number
-					u8 payload_64 = 0;
+					u64 payload_64 = 0;
 					for(int i = 0; i < 8; i = i + 1)
 					{
 						payload_64 |= ((u64)uart_get_buffer_byte()) << (i * 8);
